requirements.gpu.txt
numpy==1.17.0
alabaster==0.7.12
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
cffi==1.11.5
chardet==3.0.4
ConfigSpace==0.4.10
cryptography==2.4.2
cycler==0.10.0
Cython==0.29.13
docutils==0.14
funcy==1.11
future==0.17.1
gensim==3.7.2
gluonnlp==0.7.1
graphviz==0.8.4
hpbandster==0.7.4
idna==2.8
imagesize==1.1.0
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
kiwisolver==1.0.1
llvmlite==0.28.0
MarkupSafe==1.1.1
matplotlib==3.0.3
more-itertools==7.0.0
mxnet-cu100==1.5.0
netifaces==0.10.9
numba==0.43.0
numexpr==2.6.9
packaging==19.0
pandas==0.25.0
patsy==0.5.1
pluggy==0.9.0
py==1.8.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.3.1
pyLDAvis==2.1.2
pyOpenSSL==18.0.0
pyparsing==2.4.2
Pyro4==4.76
PySocks==1.6.8
pytest==4.4.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.20.3
scipy==1.3.0
serpent==1.28
six==1.12.0
smart-open==1.8.3
snowballstemmer==1.2.1
Sphinx==1.8.5
sphinx-gallery==0.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-websupport==1.1.0
statsmodels==0.10.1
typing==3.7.4
umap-learn==0.3.7
urllib3==1.25.3
mantichora==0.9.5
atpbar==1.0.4
